Split up code generation into more files
